MCQ question view component
