+++ src/components/Exam/index.ts
@@ -2,3 +2,4 @@ export {QuestionCard} from './QuestionCard.tsx';
 export {McqQuestion} from './McqQuestion.tsx';
 export {AddQuestion} from './AddQuestion.tsx';
 export {QuestionsListCard} from './QuestionsListCard.tsx';
+export {McqQuestionView} from './McqQuestionView.tsx';
