+++ src/pages/candidate/ExamView.css
@@ -25,7 +25,7 @@
 .radio-button {
     margin-right: 8px;
     transform: scale(1.6); /* Make the radio button larger */
-    transition: accent-color 0.3s ease;
+    transition: accent-color 0.5s ease;
 }
 
 .radio-button:hover,
