+++ src/pages/candidate/ExamView.tsx
@@ -1,5 +1,6 @@
 import { Row, Col, theme } from 'antd';
-import { CSSProperties, useState } from 'react';
+import { CSSProperties } from 'react';
+import { McqQuestionView } from '../../components';
 import './ExamView.css';
 
 import {
@@ -20,11 +21,11 @@ export const ExamViewPage = () => {
         token: { colorBgContainer, borderRadiusLG },
       } = theme.useToken();
 
-    const questionContainerStyle: CSSProperties = {
-        padding: '1px 24px',
-        background: colorBgContainer,
-        borderRadius: borderRadiusLG,
-    };
+    // const questionContainerStyle: CSSProperties = {
+    //     padding: '1px 24px',
+    //     background: colorBgContainer,
+    //     borderRadius: borderRadiusLG,
+    // };
     
     const timeStyle: CSSProperties = {
         padding: 24,
@@ -32,20 +33,20 @@ export const ExamViewPage = () => {
         borderRadius: borderRadiusLG,
     };
 
-    const [hovered, setHovered] = useState<number | null>(null);
-    const [selected, setSelected] = useState<number | null>(null);
+    // const [hovered, setHovered] = useState<number | null>(null);
+    // const [selected, setSelected] = useState<number | null>(null);
 
-    const handleMouseEnter = (index: number) => {
-        setHovered(index);
-    };
+    // const handleMouseEnter = (index: number) => {
+    //     setHovered(index);
+    // };
 
-    const handleMouseLeave = () => {
-        setHovered(null);
-    };
+    // const handleMouseLeave = () => {
+    //     setHovered(null);
+    // };
 
-    const handleSelection = (index: number) => {
-        setSelected(index);
-    };
+    // const handleSelection = (index: number) => {
+    //     setSelected(index);
+    // };
     
     return (
 
@@ -88,35 +89,15 @@ export const ExamViewPage = () => {
                 </Col>
             </Col>
             {/* ----------------Question Container----------------- */}
-            <Col span={18} pull={6}>
-                <Col span={24}>
-                    <div style={questionContainerStyle}>
-                        <p className="question-style">5. Which of the following algorithms is typically used for classification tasks in machine learning?</p>
-                        {['K-Means Clustering', 'Linear Regression', 'Support Vector Machine (SVM)', 'Principal Component Analysis (PCA)'].map((answer, index) => (
-                            <Col span={24} key={index}>
-                                <div
-                                    className={`answer-container ${hovered === index ? 'hovered' : ''} ${selected === index ? 'selected' : ''}`}
-                                    onMouseEnter={() => handleMouseEnter(index)}
-                                    onMouseLeave={handleMouseLeave}
-                                    onClick={() => handleSelection(index)}
-                                >
-                                    <input
-                                        type="radio"
-                                        id={`option${index}`}
-                                        name="mcq"
-                                        value={`option${index}`}
-                                        className={`radio-button ${hovered === index ? 'hovered' : ''}`}
-                                        checked={selected === index}
-                                        onChange={() => handleSelection(index)}
-                                    />
-                                    <label className="answer" htmlFor={`option${index}`}>{answer}</label>
-                                </div>
-                            </Col>
-                        ))}
-                    </div>
-                </Col>
-                
-            </Col>
+            <McqQuestionView
+                    question="5. Which of the following algorithms is typically used for classification tasks in machine learning?"
+                    options={[
+                        'K-Means Clustering',
+                        'Linear Regression',
+                        'Support Vector Machine (SVM)',
+                        'Principal Component Analysis (PCA)',
+                    ]}
+            />
         </Row>
         </div>
     );
